Renaming the folder
renaming the folder to filterBySteps
